Create a common dispatcher for all things notifications
ActivityLog, External Syslog, SNMP, push notifications etc.

Single point to hold logic and route the different notifications to the
different channels they should be sent to
